Os, commands added

+++ Operations.c
@@ -6,63 +6,63 @@
 *                                                     *
 * Return: none                                        *
 ******************************************************/
-void F_OR(int *r1, int *r2, int *r3)
+void F_OR(long long *r1, long long *r2, long long *r3)
 {
     *r3 = *r1 || *r2;
 }
-void F_AND(int *r1, int *r2, int *r3)
+void F_AND(long long *r1, long long *r2, long long *r3)
 {
     *r3 = *r1 && *r2;
 }
-void F_NOT(int *r1)
+void F_NOT(long long *r1)
 {
     *r1 = !*r1;
 }
-void F_XOR(int *r1, int *r2, int *r3)
+void F_XOR(long long *r1, long long *r2, long long *r3)
 {
     *r3 = *r1 ^ *r2;
 }
-void F_NAND(int *r1, int *r2, int *r3)
+void F_NAND(long long *r1, long long *r2, long long *r3)
 {
     *r3 = !(*r1 && *r2);
 }
-void F_NOR(int *r1, int *r2, int *r3)
+void F_NOR(long long *r1, long long *r2, long long *r3)
 {
     *r3 = !(*r1 || *r2);
 }
-void F_ADD(int *r1, int *r2, int *r3)
+void F_ADD(long long *r1, long long *r2, long long *r3)
 {
     *r3 = *r1 + *r2;
 }
-void F_SUB(int *r1, int *r2, int *r3)
+void F_SUB(long long *r1, long long *r2, long long *r3)
 {
     *r3 = *r1 - *r2;
 }
-void F_DIV(int *r1, int *r2, int *r3)
+void F_DIV(long long *r1, long long *r2, long long *r3)
 {
     *r3 = *r1 / *r2;
 }
-void F_MUL (int *r1, int *r2, int *r3)
+void F_MUL (long long *r1, long long *r2, long long *r3)
 {
     *r3 = *r1 * *r2;
 }
-void F_LO(int *r1, int *r2, int *r3)
+void F_LO(long long *r1, long long *r2, long long *r3)
 {
     *r3 = *r1 % *r2;
 }
-void F_INC(int *r1)
+void F_INC(long long *r1)
 {
     (*r1)++;
 }
-void F_DEC(int *r1)
+void F_DEC(long long *r1)
 {
     *r1--;
 }
-void F_NEG(int *r1)
+void F_NEG(long long *r1)
 {
     *r1 *= -1;
 }
-void F_EQ(int *r1, int *r2, int *r3)
+void F_EQ(long long *r1, long long *r2, long long *r3)
 {
     *r3 = *r1 == *r2;
 }
